Enhance order management and deployment configurations
- Add input option to recreate dev stack in deploy workflow
- Update DynamoDB schema for orders and catalog with new attributes and indexes
- Implement idempotency key handling in createOrder function
- Refactor deleteOrder to check order status before cancellation
- Improve getOrder and getOrders to enforce user permissions
- Update order status handling with version control and transaction safety
- Modify permissions to restrict order status updates
- Add validation to prevent duplicate products in order creation
- Adjust product model to enforce integer pricing

+++ src/models/product.model.js
@@ -3,7 +3,7 @@ const { z } = require('zod');
 const createProductSchema = z.object({
   name: z.string().min(1),
   description: z.string().optional(),
-  price: z.number().nonnegative(),
+  price: z.number().int().nonnegative(),
   stock: z.number().int().nonnegative(),
   category: z.string().optional(),
 });
